Update README.md
This tool not only takes the code for checking for password strength but also applies a graphic user interface (GUI) that allows the user to interact with the code when ran. SecPass takes a user’s input for the length of their desired password and creates truly randomly generated characters and allows the user to copy that password. The key piece to this tool working is utilizing the correct Module from the Standard Library. Up to this point before the secrets module was released in 2016, Cyber Security professionals had the random module to use, though this module was not random. A malicious actor could recreate the steps another completed to create their password and generate the same line of characters making this module great for practice but not truly secure. The secrets module does have random characters that are not able to be recreated because the secrets module uses the operating system itself to create these lines of characters (2024). 
	The script for this tool first imports the requisite modules to allow for specific functions to be available without needing to script and name these functions and variables. The script then sets the parameters for the window that will pop up as well as labels seen in lines 5 - 10 in image 2. Next, there is a while true function that allows the script to be ran repeatedly until the password generated meets the conditions of at least one lower case letter, one uppercase letter and at least three numbers present. Once these conditions are met it breaks out of the while true “loop.” This portion of script is from the baseline script provided by the secrets module and should an organization need more strict criteria, can be adjusted easily in lines 22 - 27. The next most important part is the ability for the user to adjust the length of the password they want. This is done in lines 18 – 21 where the variable for the slider is named and called up again as need later like in line 23. The rest of the lines are naming variables that are also creating buttons, titles, and fields for input. These can be easily adjusted such as the font or size as well as renamed should an organization deem that necessary.